backend/requirements.txt
fastapi
uvicorn
pymupdf
paddleocr
pymilvus
llama-cpp-python
paddlepaddle
